Extract binlog tests into binlog_test.go (#716 subtask 2)
Move 7 binlog subscription tests from runner_test.go into binlog_test.go:
- TestE2EBinlogSubscribingCompositeKey
- TestE2EBinlogSubscribingCompositeKeyVarchar
- TestE2EBinlogSubscribingCompositeKeyCollation
- TestE2EBinlogSubscribingCompositeKeyBinary
- TestE2EBinlogSubscribingCompositeKeyDateTime
- TestE2EBinlogSubscribingNonCompositeKey
- TestE2EBinlogSubscribingRogueValues (renamed from TestE2ERogueValues)

TestChunkerPrefetching remains in runner_test.go as it tests chunker
behavior, not binlog subscription.

Conversions:
- All table setup converted to NewTestTable (removes manual DROP/Cleanup blocks)
- All NewRunner calls converted to NewTestRunner (removes mysql.ParseDSN boilerplate)
- Varchar test seeding converted to SeedRows with UUID()
- Collation test uses deterministic CTE seeding with mixed-case prefixes
- All assert usage converted to require (fail-fast on errors)
- Added defer m.Close() to collation and binary tests for resource cleanup
- Removed unnecessary time.Sleep (watermark set synchronously by CopyChunk)
- mysql.ParseDSN eliminated entirely from binlog_test.go

runner_test.go: 3533 -> 2386 lines (-32%)
binlog_test.go: 981 lines (new)
